+++ server/src/completion.ts
@@ -3,13 +3,13 @@ import {
   CompletionItem,
   CompletionList
 } from "vscode-languageserver"
-import { clientAndObjfromUrl, parts } from "./utilities"
+import { clientAndObjfromUrl } from "./utilities"
 import { log } from "./clientManager"
 
 export async function completion(params: CompletionParams) {
   const iRole = 58 // sccmp_role_intftype in abap
   const items: CompletionItem[] = []
-  const sapIdStartPattern = /[\w\/\<]/
+  // const sapIdStartPattern = /[\w\/\<]/
   try {
     const co = await clientAndObjfromUrl(params.textDocument.uri)
     if (!co) return items
@@ -20,24 +20,24 @@ export async function completion(params: CompletionParams) {
       params.position.line + 1,
       params.position.character
     )
-    let prefix: string
+    // let prefix: string
+    const line = source.split(/\n/)[params.position.line] || ""
     rawItems.forEach(i => {
       const label = i.IDENTIFIER + (i.ROLE === iRole ? "~" : "")
       let insertText = label
-      if (label && label[0].match(/^[\/<]/)) {
-        // if the first character of the match is < or a namespace, eat it from completion text
-        if (prefix === undefined) {
-          const line = source.split(/\n/)[params.position.line]
-          let pos = params.position.character
-          while (pos > 0 && line[pos - 1].match(sapIdStartPattern)) pos--
-          // field symbol
-          if (label[0] === "<") prefix = line && line[pos]
-          // namespace
-          else [prefix] = parts(line.substr(pos), /^(\/(?:[\w]+\/)?)/)
-          // }
-          prefix = prefix || "" // prevent doing it every loop
-        }
-        if (prefix) insertText = insertText.substr(prefix.length)
+      // fix namespaces
+      const match = label.match(/^(\/\w+\/)/)
+      if (match) {
+        let len = match[1].length
+        len = i.PREFIXLENGTH >= len ? len : 1
+        if (len) insertText = insertText.substr(len)
+      }
+      // fix field-symbols
+      if (label[0] === "<") {
+        if (line[params.position.character - i.PREFIXLENGTH] === "<")
+          insertText = insertText.substr(1)
+        if (line[params.position.character] === ">")
+          insertText = insertText.substr(0, insertText.length - 1)
       }
       const item: CompletionItem = {
         label,
@@ -50,5 +50,19 @@ export async function completion(params: CompletionParams) {
   } catch (e) {
     log("Exception in completion:", e.toString()) // ignore
   }
-  return CompletionList.create(items, items.length > 10)
+  const isInComplete = (compl: CompletionItem[]) => {
+    if (compl.length > 10) return true
+    if (compl.length === 0) return false
+    // special handling for "type table of"
+    let found = false
+
+    compl.some(c => {
+      const match = c.label.match(/type( table of)?/i)
+      if (match && match[1]) found = true
+      return found || !match
+    })
+
+    return !found
+  }
+  return CompletionList.create(items, isInComplete(items))
 }